Pushing current progress
