+++ EXScreen.h
@@ -1,63 +1,48 @@
-#ifndef EX_SCREEN_H
-#define EX_SCREEN_H
+#ifndef EXSCREEN_H
+#define EXSCREEN_H
 
-// EX-Screen.h
+#include <Arduino.h>
 
-#define DEBUG
+class EXScreen {
+public:
+  /// @brief Constructor for a new EXScreen object
+  /// @param screenId The ID for this screen, should match the SCREEN() call
+  /// @param rows Number of rows valid for this screen
+  /// @param columns Number of columns valid for this screen
+  EXScreen(uint8_t screenId, uint8_t rows, uint8_t columns);
 
-uint8_t currentScreenID = 0;
+  /// @brief Get this screen's ID
+  /// @return 
+  uint8_t getId();
 
-// ALL Touch panels and wiring are DIFFERENT.  The touch screen needs to be calibrated
-// See the README files for how to run the calibration routine and
-// copy-paste results from TouchScreen_Calibr_native.ino into the lines below.
-// These settings are for the McuFriend 2.8" shield in Portrait tft.screenRotaion(0)
+  /// @brief Get the row count for this screen
+  /// @return 0 - 255
+  uint8_t getRows();
 
-//const byte XP = 7, XM = A1, YP = A2, YM = 6;
-//const int TS_LEFT=899,TS_RT=122,TS_TOP=100,TS_BOT=898;
-//const int TS_LEFT=122,TS_RT=899,TS_TOP=100,TS_BOT=898;
+  /// @brief Get the column count for this screen
+  /// @return 0 - 255
+  uint8_t getColumns();
 
-// Define some colours for the display
+  /// @brief Virtual function to override to write a row of text to the physical screen
+  virtual void writeRow(uint8_t row, uint8_t column, char *message);
 
-#define BLACK   0x0000U
-#define RED     0xF800U
-#define GREEN   0x07E0U
-#define CYAN    0x07FFU
-#define MAGENTA 0xF81FU
-#define YELLOW  0xFFE0U
-#define WHITE   0xFFFFU
-//#define YELLOW  0xE0U
-//#define WHITE   0xFFU
-//const uint16_t YELLOW = 0xFFE0U;
-//const uint16_t WHITE = 0xFFFFU;
+  /// @brief Get the first screen object
+  /// @return Pointer to the first screen object in the list
+  static EXScreen *getFirst();
 
+  /// @brief Get the next screen object
+  /// @return Pointer to the next screen object in the list
+  EXScreen *getNext();
 
-/*#if defined(ESP32)
-  #define YELLOW ~YELLOW
-  #define WHITE ~WHITE
-#endif*/
+private:
+  // Chaining screens
+  static EXScreen *_first;
+  EXScreen *_next;
+  // Screen attributes
+  uint8_t _id;
+  uint8_t _rows;
+  uint8_t _columns;
 
-// variables to indicate what needs doing to display the screen
-bool ScreenChanged[MAX_SCREENS];
-bool PrintInProgress=false;
-byte NextRowToPrint=0;
-byte NextScreenLine=0;
-
-namespace SCREEN
-{
-
-    void TFT_Startup();
-    void showmsgXY(int x, int y, int sz, const char *msg);
-    void TFT_DrawHeader();
-    void testprint(byte lines);
-    void setScreenRows(uint8_t rowId);
-    void CheckScreens();
-    void StartScreenPrint();
-    void PrintNoData();
-    //void PrintSingleLine(byte screenNo, byte screenRow);
-    void PrintALine(int Row, char * text);
-    void DisplayScreen();
-    void check_touch();
-
-}
+};
 
 #endif
